requirements.txt
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
numpy==2.0.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
